requirements.txt
numpy==1.24.2
pandas==2.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.28.2